refactor: contexts have been refactored to work directly on them and not through hooks
