+++ vite/lion/src/middleware/RolesMiddleware.jsx
@@ -1,15 +1,15 @@
 /* eslint-disable react/prop-types */
-import Unauthorized from "../pages/errors/Unauthorized.jsx";
-import { useAuth } from "../context/AuthProvider.jsx";
-import { Fragment } from "react";
-import useAES from "../hooks/useAES.jsx";
+import { Fragment, useContext } from "react";
+import useAES from "../hooks/useAES";
+import Unauthorized from "../pages/errors/Unauthorized";
+import { AuthContext } from "../context/AuthContext";
 
 export default function RolesMiddleware({
   children,
   roles,
   unauthorized = true,
 }) {
-  const { jwt } = useAuth();
+  const { jwt } = useContext(AuthContext);
   const { decode } = useAES();
 
   return jwt.idroles && !roles.includes(parseInt(decode(jwt.idroles))) ? (
